Create GPS_SIMULATED_DISTANCE_CALC
NOTE: WHEN USING THIS CODE, ONLY USE LAT/LON THAT IS DEGREES NORTH AND DEGREES WEST, DO NOT USE DEGREES EAST IT WILL BREAK.

This code is used for very base-form simulations of an RC-aircraft travelling a long distance. It uses GPS coordinates and a user inputted average speed of their aricraft to track how close the aircraft is getting to its destination, and it stops whenever its within < 1 nautical mile of distance.

This code essentially takes the distance between two points, point A and point B. Point A is a user inputted latitude AND longitude coordinate. This is named "lat_start" and "lon_start" respectively. The user will then input the point B latitude and longitude, aptly named "lat_end" and "lon_end."

Then, the user will be asked to input the average speed of the aircraft. This is a speed in miles per hour (mi/h.) After the user has inputted these 5 variables, the simulation will start.

After the simulation is done, the program will output the duration of the simulated flight in hours, minutes, seconds and milliseconds. It will also output the total distance it travelled and what the calculated average speed is.